refactor(http-client-java): consolidate TypeSpec mappers into base mappers
Merge the TypeSpec-specific mappers into the shared base mappers and relocate the remaining flavor mappers/factories for a simpler, more consistent hierarchy. No change to generated output.
