+++ .chronus/changes/http-client-java-consolidate-mappers-2026-6-24.md
@@ -1,0 +1,8 @@
+---
+changeKind: internal
+packages:
+  - "@typespec/http-client-java"
+---
+
+Simplify the mapper class hierarchy by consolidating the TypeSpec-specific mappers into
+the shared base mappers.
