+++ packages/http-client-java/generator/http-client-generator-mgmt/src/main/java/com/microsoft/typespec/http/client/generator/mgmt/mapper/FluentMapperFactory.java
@@ -3,16 +3,14 @@
 
 package com.microsoft.typespec.http.client.generator.mgmt.mapper;
 
-import com.microsoft.typespec.http.client.generator.core.mapper.ChoiceMapper;
 import com.microsoft.typespec.http.client.generator.core.mapper.ClientMethodMapper;
 import com.microsoft.typespec.http.client.generator.core.mapper.DefaultMapperFactory;
 import com.microsoft.typespec.http.client.generator.core.mapper.ExceptionMapper;
 import com.microsoft.typespec.http.client.generator.core.mapper.MethodGroupMapper;
 import com.microsoft.typespec.http.client.generator.core.mapper.ModelMapper;
+import com.microsoft.typespec.http.client.generator.core.mapper.ModelPropertyMapper;
 import com.microsoft.typespec.http.client.generator.core.mapper.ObjectMapper;
-import com.microsoft.typespec.http.client.generator.core.mapper.PrimitiveMapper;
 import com.microsoft.typespec.http.client.generator.core.mapper.ProxyMethodMapper;
-import com.microsoft.typespec.http.client.generator.core.mapper.SealedChoiceMapper;
 
 public class FluentMapperFactory extends DefaultMapperFactory {
 
@@ -42,22 +40,12 @@ public class FluentMapperFactory extends DefaultMapperFactory {
     }
 
     @Override
-    public PrimitiveMapper getPrimitiveMapper() {
-        return FluentPrimitiveMapper.getInstance();
-    }
-
-    @Override
     public ModelMapper getModelMapper() {
         return FluentModelMapper.getInstance();
     }
 
     @Override
-    public ChoiceMapper getChoiceMapper() {
-        return FluentChoiceMapper.getInstance();
-    }
-
-    @Override
-    public SealedChoiceMapper getSealedChoiceMapper() {
-        return FluentSealedChoiceMapper.getInstance();
+    public ModelPropertyMapper getModelPropertyMapper() {
+        return FluentModelPropertyMapper.getInstance();
     }
 }
