--- packages/http-client-java/generator/http-client-generator/src/main/java/com/microsoft/typespec/http/client/generator/fluent/TypeSpecFluentObjectMapper.java
@@ -1,20 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.microsoft.typespec.http.client.generator.fluent;
-
-import com.microsoft.typespec.http.client.generator.core.extension.model.codemodel.ObjectSchema;
-import com.microsoft.typespec.http.client.generator.mgmt.mapper.FluentObjectMapper;
-
-public class TypeSpecFluentObjectMapper extends FluentObjectMapper {
-    private static final TypeSpecFluentObjectMapper INSTANCE = new TypeSpecFluentObjectMapper();
-
-    public static TypeSpecFluentObjectMapper getInstance() {
-        return INSTANCE;
-    }
-
-    @Override
-    public boolean isPlainObject(ObjectSchema compositeType) {
-        return false;
-    }
-}
