+++ packages/http-client-java/generator/http-client-generator/src/main/java/com/microsoft/typespec/http/client/generator/mapper/AzureDataPlaneMapperFactory.java
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.microsoft.typespec.http.client.generator.mapper;
+
+import com.microsoft.typespec.http.client.generator.core.mapper.DefaultMapperFactory;
+import com.microsoft.typespec.http.client.generator.core.mapper.MethodGroupMapper;
+
+public class AzureDataPlaneMapperFactory extends DefaultMapperFactory {
+
+    @Override
+    public MethodGroupMapper getMethodGroupMapper() {
+        return AzureDataPlaneMethodGroupMapper.getInstance();
+    }
+}
