+++ packages/http-client-java/generator/http-client-generator/src/main/java/com/microsoft/typespec/http/client/generator/mapper/AzureDataPlaneMethodGroupMapper.java
@@ -1,23 +1,24 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.typespec.http.client.generator.core.mapper;
+package com.microsoft.typespec.http.client.generator.mapper;
 
 import com.microsoft.typespec.http.client.generator.core.extension.model.codemodel.Client;
+import com.microsoft.typespec.http.client.generator.core.mapper.MethodGroupMapper;
 import com.microsoft.typespec.http.client.generator.core.util.ClientModelUtil;
 import com.microsoft.typespec.http.client.generator.core.util.SchemaUtil;
 
 /**
- * Mapper entry point for TypeSpec-specific method group customizations.
+ * Mapper entry point for Azure data-plane method group customizations.
  */
-public class TypeSpecMethodGroupMapper extends MethodGroupMapper {
-    private static final TypeSpecMethodGroupMapper INSTANCE = new TypeSpecMethodGroupMapper();
+public class AzureDataPlaneMethodGroupMapper extends MethodGroupMapper {
+    private static final AzureDataPlaneMethodGroupMapper INSTANCE = new AzureDataPlaneMethodGroupMapper();
 
-    protected TypeSpecMethodGroupMapper() {
+    protected AzureDataPlaneMethodGroupMapper() {
         super();
     }
 
-    public static TypeSpecMethodGroupMapper getInstance() {
+    public static AzureDataPlaneMethodGroupMapper getInstance() {
         return INSTANCE;
     }
