--- packages/http-client-java/generator/http-client-generator/src/main/java/com/microsoft/typespec/http/client/generator/mapper/TypeSpecAzureVNextMapperFactory.java
@@ -1,17 +1,0 @@
-package com.microsoft.typespec.http.client.generator.mapper;
-
-import com.microsoft.typespec.http.client.generator.core.mapper.ClientMapper;
-import com.microsoft.typespec.http.client.generator.core.mapper.PrimitiveMapper;
-import com.microsoft.typespec.http.client.generator.core.mapper.azurevnext.AzureVNextMapperFactory;
-
-public class TypeSpecAzureVNextMapperFactory extends AzureVNextMapperFactory {
-    @Override
-    public ClientMapper getClientMapper() {
-        return TypeSpecClientMapper.getInstance();
-    }
-
-    @Override
-    public PrimitiveMapper getPrimitiveMapper() {
-        return TypeSpecPrimitiveMapper.getInstance();
-    }
-}
