--- packages/http-client-java/generator/http-client-generator/src/main/java/com/microsoft/typespec/http/client/generator/mapper/TypeSpecMapperFactory.java
@@ -1,34 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.microsoft.typespec.http.client.generator.mapper;
-
-import com.microsoft.typespec.http.client.generator.core.mapper.ClientMapper;
-import com.microsoft.typespec.http.client.generator.core.mapper.DefaultMapperFactory;
-import com.microsoft.typespec.http.client.generator.core.mapper.MethodGroupMapper;
-import com.microsoft.typespec.http.client.generator.core.mapper.PrimitiveMapper;
-import com.microsoft.typespec.http.client.generator.core.mapper.ProxyParameterMapper;
-import com.microsoft.typespec.http.client.generator.core.mapper.TypeSpecMethodGroupMapper;
-
-public class TypeSpecMapperFactory extends DefaultMapperFactory {
-
-    @Override
-    public ClientMapper getClientMapper() {
-        return TypeSpecClientMapper.getInstance();
-    }
-
-    @Override
-    public PrimitiveMapper getPrimitiveMapper() {
-        return TypeSpecPrimitiveMapper.getInstance();
-    }
-
-    @Override
-    public ProxyParameterMapper getProxyParameterMapper() {
-        return TypeSpecProxyParameterMapper.getInstance();
-    }
-
-    @Override
-    public MethodGroupMapper getMethodGroupMapper() {
-        return TypeSpecMethodGroupMapper.getInstance();
-    }
-}
